Add CSS bundling configuration, reorganize project
