+++ package.json
@@ -20,12 +20,16 @@
     "@rollup/plugin-commonjs": "^16.0.0",
     "@rollup/plugin-node-resolve": "^10.0.0",
     "@rollup/plugin-typescript": "^6.1.0",
+    "cssnano": "^4.1.10",
     "koa": "^2.13.0",
     "koa-send": "^5.0.1",
+    "postcss": "^8.1.10",
+    "postcss-preset-env": "^6.7.0",
     "prettier": "^2.2.0",
     "rollup": "^2.33.3",
     "rollup-plugin-hotreload": "^2.3.1",
     "rollup-plugin-livereload": "^2.0.0",
+    "rollup-plugin-postcss": "^3.1.8",
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.29.0",
