+++ src/components/Grid/Grid.css
@@ -1,13 +1,3 @@
-:root {
-  --default-color: red;
-  --default-side-length: 6vmin;
-}
-
-body {
-  width: 100vw;
-  height: 100vh;
-}
-
 .grid__parent {
   width: var(--width);
   height: var(--height);
@@ -27,10 +17,4 @@ body {
   );
   grid-auto-rows: var(--size, var(--default-side-length));
   justify-content: center;
-}
-
-.grid__square {
-  width: 100%;
-  height: 100%;
-  background: var(--color, var(--default-color));
 }
