+++ src/components/Square/Square.css
@@ -1,0 +1,5 @@
+.grid__square {
+  width: 100%;
+  height: 100%;
+  background: var(--color, var(--default-color));
+}
