+++ src/index.ts
@@ -1,5 +1,8 @@
-import Grid from './Grid'
+import Grid from './components/Grid/Grid'
 import { generateArray } from './util'
+
+import './reset.css'
+import './style.css'
 
 const colors = [['#fdc5f5', '#f7aef8', '#b388eb', '#8093f1', '#72ddf7']]
