+++ src/style.css
@@ -1,0 +1,4 @@
+body {
+  width: 100vw;
+  height: 100vh;
+}
